items: create, edit, delete, view

+++ SMCProcurement/item/routes.py
@@ -7,7 +7,7 @@ from SMCProcurement import login_manager, db
 from jinja2 import TemplateNotFound
 
 from SMCProcurement.item.forms import ItemForm
-from SMCProcurement.models import ItemCategory, Supplier, Department
+from SMCProcurement.models import ItemCategory, Supplier, Department, RequestLine
 from SMCProcurement.models.item import Item
 
 
@@ -24,7 +24,6 @@ def create_item():
     form = ItemForm()
     form.category_id.choices = [(d.id, d.name) for d in ItemCategory.query.all()]
     form.supplier_id.choices = [(d.id, d.name) for d in Supplier.query.all()]
-    form.department_id.choices = [(d.id, d.name) for d in Department.query.all()]
 
     if 'create_item' in request.form:
         item = Item(**request.form)
@@ -44,7 +43,6 @@ def edit_item(id):
     form = ItemForm(obj=item)
     form.category_id.choices = [(d.id, d.name) for d in ItemCategory.query.all()]
     form.supplier_id.choices = [(d.id, d.name) for d in Supplier.query.all()]
-    form.department_id.choices = [(d.id, d.name) for d in Department.query.all()]
 
     if 'edit_item' in request.form:
         item.update(**request.form)
@@ -62,6 +60,10 @@ def delete_item(id):
     if 'delete_item' in request.form.to_dict():
         try:
             item = db.session.query(Item).get(id)
+
+            requestLines = db.session.query(RequestLine).filter_by(item_id=id).count()
+            if requestLines > 0:
+                raise ValueError("Item is actively used in request")
             db.session.delete(item)
             db.session.commit()
